Fix IDOR in article media deletion

+++ app/Http/Controllers/Dashboard/ArticleController.php
@@ -86,18 +86,43 @@ class ArticleController extends Controller
   public function deleteImage($id, $fileName)
   {
     $article = Article::findOrFail($id);
-    if ($article->image !== 'default.jpg' && File::exists(public_path(self::IMAGE_PATH . $fileName))) {
+
+    if (in_array(Auth::user()->role->name, ['user', 'author']) && $article->user_id !== Auth::id()) {
+      abort(403);
+    }
+
+    $fileName = basename($fileName);
+
+    if (
+      $article->image === $fileName &&
+      $article->image !== 'default.jpg' &&
+      File::exists(public_path(self::IMAGE_PATH . $fileName))
+    ) {
       File::delete(public_path(self::IMAGE_PATH . $fileName));
       $article->update(['image' => 'default.jpg']);
+    } else {
+      abort(404);
     }
   }
 
   public function deleteVideo($id, $fileName)
   {
     $article = Article::findOrFail($id);
-    if (File::exists(public_path(self::VIDEO_PATH . $fileName))) {
+
+    if (in_array(Auth::user()->role->name, ['user', 'author']) && $article->user_id !== Auth::id()) {
+      abort(403);
+    }
+
+    $fileName = basename($fileName);
+
+    if (
+      $article->video === $fileName &&
+      File::exists(public_path(self::VIDEO_PATH . $fileName))
+    ) {
       File::delete(public_path(self::VIDEO_PATH . $fileName));
       $article->update(['video' => null]);
+    } else {
+      abort(404);
     }
   }
